Move formats package to pkg/formats
