+++ pkg/formats/util.go
