+++ pkg/formats/xml_test.go
